+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 - Support warning about and automatically fixing some simple uses of some 1.0 elements to their 2.0 usage styles.
-- Support warning about and automatically fixing element declarations that use `<content>` to use `<slot>` instead.
+- Support warning about and automatically fixing element declarations that use `<content>` to use `<slot>` instead. This is a breaking change to an element,
+so it is not on by default.
 
 ## [2.1.0] - 2017-10-13
 - Warn for old-style @apply without parentheses, and var() with a fallback value of a bare css variable.